requirements_live.txt
django
twilio
Pillow==8.0.0
razorpay
sslcommerz-lib
openai==0.28.0
paypalrestsdk==1.5.0
requests
instamojo-wrapper
stripe
django-ckeditor==6.3.2
django_user_agents
django-imagekit
urllib3==1.25.11
whitenoise==6.5.0
mysqlclient==2.1.1